Changed the result table
Now the result table only displays candy, report, and reward.  It does not display the ref player or the ref report.

The ref report can be inferred from the report and the reward.
The ref player cannot be inferred.  It needs to be discussed whether it is important to display the ref player or not.