Changed the updateActions function
Whenever the server receives a report from a player, the server sends confirmation messages to all clients to inform them of a new report or to inform the sender that the report has been received.  So I modified the updateActions method on the client side.  Whenever it receives a message, it expects the message to contain the index of a player who sent a report to the server.  The client updates his record of which of the other players have confirmed their reports.